Rename endpoints to @tus-upload and @tus-replace
