+++ docs/source/_json/tusupload_options.req
@@ -1,3 +1,3 @@
-OPTIONS /plone/folder/@upload HTTP/1.1
+OPTIONS /plone/folder/@tus-upload HTTP/1.1
 Accept: application/json
 Authorization: Basic YWRtaW46c2VjcmV0
