+++ docs/source/_json/tusupload_patch_finalized.req
@@ -1,4 +1,4 @@
-PATCH /plone/folder/@upload/032803b64ad746b3ab46d9223ea3d90f HTTP/1.1
+PATCH /plone/folder/@tus-upload/032803b64ad746b3ab46d9223ea3d90f HTTP/1.1
 Accept: application/json
 Authorization: Basic YWRtaW46c2VjcmV0
 Tus-Resumable: 1.0.0
