+++ docs/source/_json/tusupload_post.req
@@ -1,4 +1,4 @@
-POST /plone/folder/@upload HTTP/1.1
+POST /plone/folder/@tus-upload HTTP/1.1
 Accept: application/json
 Authorization: Basic YWRtaW46c2VjcmV0
 Tus-Resumable: 1.0.0
